+++ src/zabbix_enums/common/prototype.py
@@ -1,5 +1,5 @@
 from .trigger import TriggerPriority, TrigerStatus, TriggerType, TriggerRecoveryMode, TriggerCorrelationMode, TriggerManualClose
-from .._common import PrototypeDiscover
+from . import _PrototypeDiscover
 
 TriggerPrototypePriority = TriggerPriority
 TriggerPrototypeStatus = TrigerStatus
@@ -7,4 +7,4 @@ TriggerPrototypeType = TriggerType
 TriggerPrototypeRecoveryMode = TriggerRecoveryMode
 TriggerPrototypeCorrelationMode = TriggerCorrelationMode
 TriggerPrototypeManualClose = TriggerManualClose
-TriggerPrototypeDiscover = PrototypeDiscover
+TriggerPrototypeDiscover = _PrototypeDiscover
