+++ src/zabbix_enums/specific/z50/dashboard.py
@@ -1,0 +1,25 @@
+from enum import Enum
+
+
+class DashboardWidgetType(str, Enum):
+    ACTION_LOG = "actionlog"
+    CLOCK = "clock"
+    DATA_OVERVIEW = "dataover"
+    DISCOVERY_STATUS = "discovery"
+    FAVORITE_GRAPHS = "favgraphs"
+    FAVORITE_MAPS = "favmaps"
+    FAVORITE_SCREENS = "favscreens"
+    GRAPH_CLASSIC = "graph"
+    GRAPH_PROTOTYPE = "graphprototype"
+    HOST_AVAILABILITY = "hostavail"
+    MAP = "map"
+    MAP_NAVIGATION_TREE = "navtree"
+    PLAIN_TEXT = "plaintext"
+    PROBLEM_HOSTS = "problemhosts"
+    PROBLEMS = "problems"
+    PROBLEMS_BY_SEVERITY = "problemsbysv"
+    GRAPH = "svggraph"
+    SYSTEM_INFORMATION = "systeminfo"
+    TRIGGER_OVERVIEW = "trigover"
+    URL = "url"
+    WEB_MONITORING = "web"
