+++ src/zabbix_enums/specific/z50/item.py
@@ -1,0 +1,53 @@
+from enum import IntEnum
+
+
+class ItemType(IntEnum):
+    ZABBIX_AGENT = 0
+    ZABBIX_TRAPPER = 2
+    SIMPLE_CHECK = 3
+    ZABBIX_INTERNAL = 5
+    ZABBIX_AGENT_ACTIVE = 7
+    ZABBIX_AGGREGATE = 8
+    WEB_ITEM = 9
+    EXTERNAL_CHECK = 10
+    DATABASE_MONITOR = 11
+    IPMI_AGENT = 12
+    SSH_AGENT = 13
+    TELNET_AGENT = 14
+    CALCULATED = 15
+    JMX_AGENT = 16
+    SNMP_TRAP = 17
+    DEPENDENT_ITEM = 18
+    HTTP_AGENT = 19
+    SNMP_AGENT = 20
+
+
+
+
+
+class ItemPreprocessingType(IntEnum):
+    CUSTOM_MULTIPLIER = 1
+    RIGHT_TRIM = 2
+    LEFT_TRIM = 3
+    TRIM = 4
+    REGULAR_EXPRESSION = 5
+    BOOLEAN_TO_DECIMAL = 6
+    OCTAL_TO_DECIMAL = 7
+    HEXADECIMAL_TO_DECIMAL = 8
+    SIMPLE_CHANGE = 9
+    CHANGE_PER_SECOND = 10
+    XML_XPATH = 11
+    JSONPATH = 12
+    IN_RANGE = 13
+    MATCHES_REGULAR_EXPRESSION = 14
+    DOES_NOT_MATCH_REGULAR_EXPRESSION = 15
+    CHECK_FOR_ERROR_IN_JSON = 16
+    CHECK_FOR_ERROR_IN_XML = 17
+    CHECK_FOR_ERROR_USING_REGULAR_EXPRESSION = 18
+    DISCARD_UNCHANGED = 19
+    DISCARD_UNCHANGED_WITH_HEARTBEAT = 20
+    JAVASCRIPT = 21
+    PROMETHEUS_PATTERN = 22
+    PROMETHEUS_TO_JSON = 23
+    CSV_TO_JSON = 24
+    REPLACE = 25
