+++ src/zabbix_enums/specific/z50/user_macro.py
@@ -1,0 +1,6 @@
+from enum import IntEnum
+
+
+class MacroType(IntEnum):
+    TEXT = 0
+    SECRET = 1
