+++ src/zabbix_enums/specific/z54/__init__.py
@@ -1,0 +1,4 @@
+from .audit_log import *
+from .dashboard import *
+from .item import *
+from .user_macro import *
