+++ src/zabbix_enums/specific/z54/user_macro.py
@@ -1,0 +1,7 @@
+from enum import IntEnum
+
+
+class MacroType(IntEnum):
+    TEXT = 0
+    SECRET = 1
+    VAULT = 2
